moved attributes to another file
